Update variation.process.jquery.js
I'm not sure if I can justify this PR without the other changes I've made and added to my site's functions.php. However, this small tweak takes into account the scenario where a user has placeholder text in the quantity select box (it is required that the option be set with the attribute `value="0"`). Prior to this change, the price update would show NaN as a new total price. After this change, the update looks to see if the value is 0. Otherwise, it takes the value of the text.